Restrict wallet creation with config flag
KDF startup throws an error if the `allow_registrations` config value is false and the seed is unknown (no name provided) or if a name is provided and it has not been registered before.

This is used to emulate the sign-in vs register behaviour a typical auth service would provide.