stream orderbook reactively
this streams out the orderbook to subbed clients as the orders arrive to mm2.

the model used for this case is a different streamer for each coin pair requested for streaming, so (btc, ltc) has a streamer, (btc, kmd) has another, and (ltc, btc) also has another.
we might want to combine the first and last cases (btc, ltc) & (ltc, btc) into a single streamer? but then we would want the streamer to have a deterministic name/id (based on lexagraphical order), or tolerate different names for the same streamer (might be confusing and also harder to impl).