Prior to this fix, --all-features builds will run this test logic because of the misuse of any() instead of all()
This pattern of introducing conditional logic within a function to *test the function itself* is particularly atrocious and must be avoided. We could further clean this up by removing these conditional blocks. This commit at least ensures that a `--all-features` build will not have test-specific logic running in prod.

Please carefully review the commented blocks in taker_restart.rs because this logic wasn't well analyzed at the time of this commit.