fix: actually (auto) remove the client when they disconnect
this commit makes it so that the client deletion logic is handled by the
streaming manager when initializing the client. the return type of
client initialization has an extra field that will run the client
delection logic on Drop.

this fixes the problem with clients never getting removed as when the clients disconnect
and the body stream is no longer drived further thus will not hit the `remove_client` call.